refactor: move types to the model
